// src/generated/ResultKind.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ResultKind = "app" | "file" | "folder" | "clipboard" | "snippet" | "quicklink" | "emoji" | "calculation" | "dateTime" | "password" | "url" | "webSearch" | "system" | "color" | "permissions";